Add dummy GoCard as a provider
This is the start of using a loosely couple provider API so new transit
providers can easily be added to the app. Ideally there'll be no
coupling between the providers and the rest of the app other than
sending messages through Redux. This includes injecting their own
reducers and sagas.