Make GoCard provider work
GoCard is the public transport card system used in Southeast Queensland,
Australia. It covers Brisbane, the Sunshine Coast, and the Gold Coast.
This is the first provider in the system. User details can't be changed
or added yet and this is work that still needs to happen.

I've been trying to think about how I can stop loading only once all the
different providers are loaded. Considering that I've currently only got
one provider this isn't much of an issue now. I've also decided that
having loading finish once the first card arrives will create a snappier
experience for users and other cards can pop into the list as they're
loaded. We'll see how this goes and reconsider if needed.

I'm adding this first because it's most useful to me.